requirements.txt
pandas
numpy
pyarrow
python-dotenv
scikit-learn
fuzzywuzzy
python-Levenshtein
